ARTEMIS-5534 use Objects.requireNonNullElse... where sensible
The code-base is littered with uses of Java's ternary operator to check
for null and fallback to a default value. Many of these can be replaced
with either `Objects.requireNonNullElse` or
`Objects.requireNonNullElseGet`. The latter is used where appropriate
to essentially match the short-circuit semantics provided by the
ternary operator.